Split generic diagnostics tests
