+++ tests/generic_diagnostics.rs
@@ -26,1602 +26,13 @@ fn frontend_errors(source: &str) -> Vec<Diagnostic> {
     }
 }
 
-#[test]
-fn nongeneric_method_explicit_type_args_are_error() {
-    let errors = typecheck_errors(
-        r#"
-Box: {
-    value: i32
-}
-
-Box.get = (self: Box) i32 {
-    self.value
-}
-
-main = () i32 {
-    box = Box { value: 1 }
-    box.get<i32>()
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("non-generic method `Box.get` does not accept type arguments")),
-        "expected non-generic method type-argument diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_explicit_type_arg_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Box.get<T> = (self: Box<T>) T {
-    self.value
-}
-
-main = () i32 {
-    box = Box<i32> { value: 1 }
-    box.get<i32, str>()
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic method `Box.get` expects 1 type arguments, found 2")),
-        "expected generic method arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_inference_failure_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box: {
-    value: i32
-}
-
-Box.make<T> = (self: Box) T {
-    self.value
-}
-
-main = () i32 {
-    box = Box { value: 1 }
-    box.make()
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("cannot infer type argument `T` for generic method `Box.make`")),
-        "expected generic method inference diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_argument_arity_uses_method_diagnostic() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Box.get<T> = (self: Box<T>) T {
-    self.value
-}
-
-main = () i32 {
-    box = Box<i32> { value: 1 }
-    box.get(2)
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("method `Box.get` expects 1 arguments, found 2")),
-        "expected generic method arity diagnostic to name method kind, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_function_explicit_type_arg_arity_does_not_emit_inference_followup() {
-    let errors = typecheck_errors(
-        r#"
-pick<T, U> = (value: T) T {
-    value
-}
-
-main = () i32 {
-    pick<i32>(1)
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic function `pick` expects 2 type arguments, found 1")),
-        "expected explicit generic arity diagnostic, got {errors:?}"
-    );
-    assert!(
-        errors
-            .iter()
-            .all(|d| !d.message.contains("cannot infer type argument")),
-        "explicit generic arity failure should not also report inference, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_explicit_type_arg_arity_does_not_emit_inference_followup() {
-    let errors = typecheck_errors(
-        r#"
-Box: {
-    value: i32
-}
-
-Box.pick<T, U> = (self: Box, value: T) T {
-    value
-}
-
-main = () i32 {
-    box = Box { value: 1 }
-    box.pick<i32>(1)
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic method `Box.pick` expects 2 type arguments, found 1")),
-        "expected explicit generic method arity diagnostic, got {errors:?}"
-    );
-    assert!(
-        errors
-            .iter()
-            .all(|d| !d.message.contains("cannot infer type argument")),
-        "explicit generic method arity failure should not also report inference, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_function_explicit_type_arg_arity_does_not_emit_argument_followup() {
-    let errors = typecheck_errors(
-        r#"
-take_second<T, U> = (value: U) U {
-    value
-}
-
-main = () i32 {
-    take_second<i32>(1)
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic function `take_second` expects 2 type arguments, found 1")),
-        "expected explicit generic arity diagnostic, got {errors:?}"
-    );
-    assert!(
-        errors.iter().all(|d| !d.message.contains("argument 1")),
-        "explicit generic arity failure should not also report argument mismatch, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_explicit_type_arg_arity_does_not_emit_argument_followup() {
-    let errors = typecheck_errors(
-        r#"
-Box: {
-    value: i32
-}
-
-Box.take_second<T, U> = (self: Box, value: U) U {
-    value
-}
-
-main = () i32 {
-    box = Box { value: 1 }
-    box.take_second<i32>(1)
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic method `Box.take_second` expects 2 type arguments, found 1")),
-        "expected explicit generic method arity diagnostic, got {errors:?}"
-    );
-    assert!(
-        errors.iter().all(|d| !d.message.contains("argument 2")),
-        "explicit generic method arity failure should not also report argument mismatch, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_function_inference_conflict_is_error() {
-    let errors = typecheck_errors(
-        r#"
-choose<T> = (left: T, right: T) T {
-    left
-}
-
-main = () i32 {
-    value = choose(1, "bad")
-    value
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "conflicting inferred type argument `T` for generic function `choose`: inferred `i32` and `str`"
-        )),
-        "expected generic function inference conflict diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_function_inference_conflict_through_function_type_is_error() {
-    let errors = typecheck_errors(
-        r#"
-choose_with<T> = (left: T, mapper: (T) T) T {
-    left
-}
-
-main = () i32 {
-    mapper = (value: str) str {
-        value
-    }
-    choose_with(1, mapper)
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "conflicting inferred type argument `T` for generic function `choose_with`: inferred `i32` and `str`"
-        )),
-        "expected generic function function-type inference conflict diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_function_inference_conflict_through_array_type_is_error() {
-    let errors = typecheck_errors(
-        r#"
-choose_array<T> = (left: T, items: [T; 1]) T {
-    left
-}
-
-main = () i32 {
-    items = ["bad"]
-    choose_array(1, items)
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "conflicting inferred type argument `T` for generic function `choose_array`: inferred `i32` and `str`"
-        )),
-        "expected generic function array-type inference conflict diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_function_inference_conflict_through_raw_pointer_type_is_error() {
-    let errors = typecheck_errors(
-        r#"
-choose_raw<T> = (left: T, ptr: RawPtr<T>) T {
-    left
-}
-
-main = () i32 {
-    ptr = cast("bad", RawPtr<str>)
-    choose_raw(1, ptr)
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "conflicting inferred type argument `T` for generic function `choose_raw`: inferred `i32` and `str`"
-        )),
-        "expected generic function raw-pointer inference conflict diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_inference_conflict_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Box.choose<T> = (self: Box<T>, other: T) T {
-    self.value
-}
-
-main = () i32 {
-    box = Box<i32> { value: 1 }
-    box.choose("bad")
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "conflicting inferred type argument `T` for generic method `Box.choose`: inferred `i32` and `str`"
-        )),
-        "expected generic method inference conflict diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_inference_conflict_from_receiver_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Box.replace<T> = (self: Box<T>, value: T) T {
-    value
-}
-
-main = () i32 {
-    box = Box<i32> { value: 1 }
-    box.replace("bad")
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "conflicting inferred type argument `T` for generic method `Box.replace`: inferred `i32` and `str`"
-        )),
-        "expected generic method receiver inference conflict diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_inference_conflict_through_function_type_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Box.choose_with<T> = (self: Box<T>, mapper: (T) T) T {
-    self.value
-}
-
-main = () i32 {
-    box = Box<i32> { value: 1 }
-    mapper = (value: str) str {
-        value
-    }
-    box.choose_with(mapper)
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "conflicting inferred type argument `T` for generic method `Box.choose_with`: inferred `i32` and `str`"
-        )),
-        "expected generic method function-type inference conflict diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_inference_conflict_through_array_type_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Box.choose_array<T> = (self: Box<T>, items: [T; 1]) T {
-    self.value
-}
-
-main = () i32 {
-    box = Box<i32> { value: 1 }
-    items = ["bad"]
-    box.choose_array(items)
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "conflicting inferred type argument `T` for generic method `Box.choose_array`: inferred `i32` and `str`"
-        )),
-        "expected generic method array-type inference conflict diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_inference_conflict_through_raw_pointer_type_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Box.choose_raw<T> = (self: Box<T>, ptr: RawPtr<T>) T {
-    self.value
-}
-
-main = () i32 {
-    box = Box<i32> { value: 1 }
-    ptr = cast("bad", RawPtr<str>)
-    box.choose_raw(ptr)
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "conflicting inferred type argument `T` for generic method `Box.choose_raw`: inferred `i32` and `str`"
-        )),
-        "expected generic method raw-pointer inference conflict diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_inference_conflict_through_slice_type_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Box.choose_slice<T> = (self: Box<T>, items: Slice<T>) T {
-    self.value
-}
-
-main = () i32 {
-    box = Box<i32> { value: 1 }
-    items = cast("bad", Slice<str>)
-    box.choose_slice(items)
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d.message.contains(
-            "conflicting inferred type argument `T` for generic method `Box.choose_slice`: inferred `i32` and `str`"
-        )),
-        "expected generic method slice inference conflict diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_function_type_arg_annotation_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-identity<T> = (value: T) T {
-    value
-}
-
-main = () i32 {
-    box = Box<i32> { value: 1 }
-    bad = identity<Box<i32, str>>(box)
-    bad.value
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 2")),
-        "expected generic function type-argument annotation arity diagnostic, got {errors:?}"
-    );
-    assert!(
-        errors.iter().all(|d| !d.message.contains("argument 1")),
-        "malformed generic function type argument should not also report argument mismatch, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_type_arg_annotation_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Holder: {
-    value: i32
-}
-
-Holder.wrap<T> = (self: Holder, value: T) T {
-    value
-}
-
-main = () i32 {
-    holder = Holder { value: 1 }
-    box = Box<i32> { value: 1 }
-    bad = holder.wrap<Box<i32, str>>(box)
-    bad.value
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 2")),
-        "expected generic method type-argument annotation arity diagnostic, got {errors:?}"
-    );
-    assert!(
-        errors.iter().all(|d| !d.message.contains("argument 2")),
-        "malformed generic method type argument should not also report argument mismatch, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_type_arg_annotation_without_type_args_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Holder: {
-    value: i32
-}
-
-Holder.wrap<T> = (self: Holder, value: T) T {
-    value
-}
-
-main = () i32 {
-    holder = Holder { value: 1 }
-    box = Box<i32> { value: 1 }
-    bad = holder.wrap<Box>(box)
-    bad.value
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 0")),
-        "expected generic method type-argument annotation without args diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn closure_param_annotation_type_arg_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-main = () i32 {
-    f = (box: Box<i32, str>) i32 {
-        0
-    }
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 2")),
-        "expected closure parameter generic annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn closure_return_annotation_without_type_args_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-main = () i32 {
-    f = () Box {
-        Box<i32> { value: 1 }
-    }
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 0")),
-        "expected closure return generic annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn cast_target_annotation_type_arg_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-main = () i32 {
-    box = Box<i32> { value: 1 }
-    bad = cast(box, Box<i32, str>)
-    bad.value
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 2")),
-        "expected cast target generic annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn as_cast_target_annotation_without_type_args_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-main = () i32 {
-    box = Box<i32> { value: 1 }
-    bad = box as Box
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 0")),
-        "expected as-cast target generic annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_struct_type_arg_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-main = () i32 {
-    box = Box<i32, str> { value: 1 }
-    box.value
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 2")),
-        "expected generic struct arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_enum_type_arg_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Option<T>:
-    None,
-    Some(T)
-
-main = () i32 {
-    value = Option<i32, str>.Some(1)
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic enum `Option` expects 1 type arguments, found 2")),
-        "expected generic enum arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_struct_annotation_type_arg_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-read = (box: Box<i32, str>) i32 {
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 2")),
-        "expected generic struct annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_enum_annotation_type_arg_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Option<T>:
-    None,
-    Some(T)
-
-read = (value: Option<i32, str>) i32 {
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic enum `Option` expects 1 type arguments, found 2")),
-        "expected generic enum annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_struct_annotation_without_type_args_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-read = (box: Box) i32 {
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 0")),
-        "expected unspecialized generic struct annotation diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_enum_annotation_without_type_args_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Option<T>:
-    None,
-    Some(T)
-
-read = (value: Option) i32 {
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic enum `Option` expects 1 type arguments, found 0")),
-        "expected unspecialized generic enum annotation diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_struct_local_annotation_type_arg_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-main = () i32 {
-    box: Box<i32, str> = Box<i32> { value: 1 }
-    box.value
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 2")),
-        "expected generic struct local annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_struct_local_annotation_without_type_args_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-main = () i32 {
-    box: Box = Box<i32> { value: 1 }
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 0")),
-        "expected unspecialized generic struct local annotation diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_enum_local_annotation_type_arg_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Option<T>:
-    None,
-    Some(T)
-
-main = () i32 {
-    value: Option<i32, str> = Option<i32>.Some(1)
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic enum `Option` expects 1 type arguments, found 2")),
-        "expected generic enum local annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_enum_local_annotation_without_type_args_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Option<T>:
-    None,
-    Some(T)
-
-main = () i32 {
-    value: Option = Option<i32>.Some(1)
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic enum `Option` expects 1 type arguments, found 0")),
-        "expected unspecialized generic enum local annotation diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn nested_generic_annotation_inner_type_arg_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Option<T>:
-    None,
-    Some(T)
-
-read = (box: Box<Option<i32, str>>) i32 {
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic enum `Option` expects 1 type arguments, found 2")),
-        "expected nested generic annotation inner arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn nested_generic_instantiation_inner_type_arg_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Option<T>:
-    None,
-    Some(T)
-
-main = () i32 {
-    value = Box<Option<i32, str>> { value: Option<i32>.Some(1) }
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic enum `Option` expects 1 type arguments, found 2")),
-        "expected nested generic instantiation inner arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn function_type_parameter_annotation_type_arg_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-call = (f: (Box<i32, str>) i32) i32 {
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 2")),
-        "expected function type parameter generic annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn function_type_return_annotation_without_type_args_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-factory = () () Box {
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 0")),
-        "expected function type return generic annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn pointer_type_inner_generic_annotation_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-read = (ptr: Ptr<Box<i32, str>>) i32 {
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 2")),
-        "expected pointer inner generic annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn slice_type_inner_generic_annotation_without_type_args_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-read = (slice: Slice<Box>) i32 {
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 0")),
-        "expected slice inner generic annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn array_type_inner_generic_annotation_arity_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-read = (items: [Box<i32, str>; 1]) i32 {
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic struct `Box` expects 1 type arguments, found 2")),
-        "expected array inner generic annotation arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_struct_behavior_bound_failure_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-Point: {
-    x: i32
-}
-
-Box<T: Json>: {
-    value: T
-}
-
-main = () i32 {
-    point = Point { x: 1 }
-    box = Box<Point> { value: point }
-    box.value.x
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json` required by `T`")),
-        "expected generic struct bound diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_enum_behavior_bound_failure_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-Point: {
-    x: i32
-}
-
-Option<T: Json>:
-    None,
-    Some(T)
-
-main = () i32 {
-    point = Point { x: 1 }
-    value = Option<Point>.Some(point)
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json` required by `T`")),
-        "expected generic enum bound diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_struct_annotation_bound_failure_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-Point: {
-    x: i32
-}
-
-Box<T: Json>: {
-    value: T
-}
-
-read = (box: Box<Point>) i32 {
-    box.value.x
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json` required by `T`")),
-        "expected generic struct annotation bound diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_enum_annotation_bound_failure_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-Point: {
-    x: i32
-}
-
-Option<T: Json>:
-    None,
-    Some(T)
-
-read = (value: Option<Point>) i32 {
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json` required by `T`")),
-        "expected generic enum annotation bound diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_struct_local_annotation_bound_failure_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-Point: {
-    x: i32
-}
-
-Box<T: Json>: {
-    value: T
-}
-
-main = () i32 {
-    point = Point { x: 1 }
-    box: Box<Point> = Box<Point> { value: point }
-    box.value.x
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json` required by `T`")),
-        "expected generic struct local annotation bound diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_enum_local_annotation_bound_failure_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-Point: {
-    x: i32
-}
-
-Option<T: Json>:
-    None,
-    Some(T)
-
-main = () i32 {
-    point = Point { x: 1 }
-    value: Option<Point> = Option<Point>.Some(point)
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json` required by `T`")),
-        "expected generic enum local annotation bound diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_function_behavior_bound_failure_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-Point: {
-    x: i32
-}
-
-encode<T: Json> = (value: T) str {
-    value.encode()
-}
-
-main = () i32 {
-    point = Point { x: 1 }
-    text = encode(point)
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json` required by `T`")),
-        "expected generic function bound diagnostic, got {errors:?}"
-    );
-    assert!(
-        errors
-            .iter()
-            .all(|d| !d.message.contains("has no method `encode`")),
-        "generic function bound failure should not also specialize body method errors, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_behavior_bound_unknown_method_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Json<T>: behavior {
-    encode: (Self) T
-}
-
-Point: {
-    x: i32
-}
-
-Point.implements(Json<Point>) {
-    encode = (value: Point) Point {
-        value
-    }
-}
-
-decode<T: Json<T>> = (value: T) T {
-    value.serialize()
-}
-
-main = () i32 {
-    point = Point { x: 1 }
-    decoded = decode(point)
-    decoded.x
-}
-"#,
-    );
-
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("type `Point` has no method `serialize`")),
-        "expected unknown method diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_method_behavior_bound_failure_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-Point: {
-    x: i32
-}
-
-Holder: {
-    value: i32
-}
-
-Holder.wrap<T: Json> = (self: Holder, value: T) T {
-    value
-}
-
-main = () i32 {
-    holder = Holder { value: 1 }
-    point = Point { x: 1 }
-    bad = holder.wrap(point)
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json` required by `T`")),
-        "expected generic method bound diagnostic, got {errors:?}"
-    );
-    assert!(
-        errors
-            .iter()
-            .all(|d| !d.message.contains("has no method `encode`")),
-        "generic method bound failure should not also specialize body method errors, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_receiver_method_behavior_bound_failure_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-Point: {
-    x: i32
-}
-
-Box<T>: {
-    value: T
-}
-
-Box.map<U: Json> = (self: Box<i32>, value: U) U {
-    value
-}
-
-main = () i32 {
-    box = Box<i32> { value: 1 }
-    point = Point { x: 1 }
-    bad = box.map(point)
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json` required by `U`")),
-        "expected generic receiver method bound diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn generic_ufc_function_behavior_bound_failure_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Json: behavior {
-    encode: (Self) str
-}
-
-Point: {
-    x: i32
-}
-
-as_json<T: Json> = (value: T) str {
-    value.encode()
-}
-
-main = () i32 {
-    point = Point { x: 1 }
-    text = point.as_json()
-    0
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("type `Point` does not implement behavior `Json` required by `T`")),
-        "expected generic UFC function bound diagnostic, got {errors:?}"
-    );
-    assert!(
-        errors
-            .iter()
-            .all(|d| !d.message.contains("has no method `encode`")),
-        "generic UFC bound failure should not also specialize body method errors, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_impl_for_unknown_type_is_error() {
-    let errors = frontend_errors(
-        r#"
-Json: behavior {
-    to_json: (Self) str
-}
-
-Missing.implements(Json) {
-    to_json = (value: Missing) str {
-        "missing"
-    }
-}
-"#,
-    );
-
-    assert!(
-        errors
-            .iter()
-            .any(|d| d.message.contains("unknown type symbol 'Missing'")),
-        "expected unknown behavior impl target diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_impl_for_unspecialized_generic_type_is_error() {
-    let errors = frontend_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-Box.implements(Json) {
-    to_json = (value: Box) str {
-        "box"
-    }
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic type `Box` expects 1 type arguments, found 0")),
-        "expected generic impl target arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_requires_unspecialized_generic_type_is_error() {
-    let errors = frontend_errors(
-        r#"
-Box<T>: {
-    value: T
-}
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-Box.requires(Json)
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| d
-            .message
-            .contains("generic type `Box` expects 1 type arguments, found 0")),
-        "expected generic requires target arity diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_impl_extra_method_is_error() {
-    let errors = typecheck_errors(
-        r#"
-Point: {
-    x: i32
-}
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-Point.implements(Json) {
-    to_json = (value: Point) str {
-        "point"
-    }
-
-    extra = (value: Point) str {
-        "extra"
-    }
-}
-"#,
-    );
-
-    assert!(
-        errors.iter().any(|d| {
-            d.message
-                .contains("method `extra` is not declared by behavior `Json`")
-        }),
-        "expected extra behavior impl method diagnostic, got {errors:?}"
-    );
-}
-
-#[test]
-fn behavior_impl_duplicate_method_is_error() {
-    let errors = frontend_errors(
-        r#"
-Point: {
-    x: i32
-}
-
-Json: behavior {
-    to_json: (Self) str
-}
-
-Point.implements(Json) {
-    to_json = (value: Point) str {
-        "point"
-    }
-
-    to_json = (value: Point) str {
-        "point again"
-    }
-}
-"#,
-    );
-
-    assert!(
-        errors
-            .iter()
-            .any(|d| { d.message.contains("duplicate value symbol 'Point.to_json'") }),
-        "expected duplicate behavior impl method diagnostic, got {errors:?}"
-    );
-}
+#[path = "generic_diagnostics/annotations.rs"]
+mod annotations;
+#[path = "generic_diagnostics/behavior_impls.rs"]
+mod behavior_impls;
+#[path = "generic_diagnostics/bounds.rs"]
+mod bounds;
+#[path = "generic_diagnostics/inference_conflicts.rs"]
+mod inference_conflicts;
+#[path = "generic_diagnostics/method_type_args.rs"]
+mod method_type_args;
